+++ plone/recipe/codeanalysis/pep3101.py
@@ -8,6 +8,11 @@ def code_analysis_pep3101(options):
     sys.stdout.write('PEP 3101')
     sys.stdout.flush()
 
+    # XXX: advice on usage of the right option
+    if options['string-formatting'] != 'False':
+        sys.stdout.write('destring-formatting option is deprecated; '
+                         'use pep3101 instead.')
+
     files = _find_files(options, '.*\.py')
     if not files:
         print('     [\033[00;32m OK \033[0m]')